% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Champs22Teams.R
\docType{data}
\name{Champs22Teams}
\alias{Champs22Teams}
\title{Valorant Champions Tour 2022 Champions Istanbul Team Data}
\format{
A data frame with 86 rows and 15 columns
\describe{
   \item{matchId}{A unique ID given to each match}
   \item{map}{Name of the map the match was played on}
   \item{teamOneId}{Lists the unique team ID for the team listed as the
   first team on the match page}
   \item{teamOneName}{Name of the team listed as the first team on the match
   page}
   \item{teamTwoId}{Lists the unique team ID for the team listed as the
   second team on the match page}
   \item{teamTwoName}{Name of the team listed as the second team on the match
   page}
   \item{teamOnePlayerId}{A character string of the unique player ID for
   each player on the team listed as the first team on the match page}
   \item{teamTwoPlayerId}{A character string of the unique player ID for
   each player on the team listed as the second team on the match page}
   \item{teamOnePlayerIgn}{A character string of the in game name for
   each player on the team listed as the first team on the match page}
   \item{teamTwoPlayerIgn}{A character string of the in game name for
   each player on the team listed as the second team on the match page}
   \item{teamOneAgents}{A character string of the agents played by each
   player on the team listed as the first team on the match page}
   \item{teamTwoAgents}{A character string of the agents played by each
   player on the team listed as the second team on the match page}
   \item{teamOneAgentRoles}{A character string of the role of the agent
   played by each player on the team listed as the first team on the match
   page}
   \item{teamTwoAgentRoles}{A character string of the role of the agent
   played by each player on the team listed as the second team on the match
   page}
   \item{winningTeamId}{Lists the unique ID of the team that won the match}
}
}
\source{
\url{https://www.rib.gg/}
}
\usage{
Champs22Teams
}
\description{
Data from all 2022 Champions Istanbul group and playoff stage matches
including team players, compositions, and roles. This data set also
contains Ids that tie this data frame to other data frames in the set.
}
\keyword{datasets}
