% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_impact.R
\name{get_impact}
\alias{get_impact}
\title{Generate impact ratings for players}
\usage{
get_impact(rpsdf, namedf)
}
\arguments{
\item{rpsdf}{The round player stats data frame}

\item{namedf}{The data frame generated by id_to_name()}
}
\value{
A data frame with 5 columns
}
\description{
Generates a data frame of calculated impact value for each player and lists
their IGN, team, total impact, and matches played.
}
\examples{
impactValues <- get_impact(Champs22RPS, champsNames)
head(impactValues)
}
\references{
\url{
https://www.rib.gg/article/measuring-impact-value-most-impactful-players-of-vct-heading-to-champions}
}
