% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_to_name.R
\name{id_to_name}
\alias{id_to_name}
\title{Link Team and Player IDs with Names}
\usage{
id_to_name(teamsdf)
}
\arguments{
\item{teamsdf}{The teams data frame}
}
\value{
A data frame with 4 columns
}
\description{
Separates and cleans the TeamId and PlayerIgn character lists to connect them
to their unique ID. Outputs a data frame with the player's IGN and unique ID
and their respective team and unique team ID.
}
\examples{
champsNames <- id_to_name(Champs22Teams)
head(champsNames)
}
