% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rib.R
\name{read_rib}
\alias{read_rib}
\title{Read in a scraped RiB data set}
\usage{
read_rib(xl)
}
\arguments{
\item{xl}{The filepath of the excel file}
}
\value{
A large list of 10 data frames
}
\description{
Reads in the excel file of the scraped RiB data and outputs a list of the 10
sheets read in as data frames.
}
\details{
I recommend running list2env(test, .GlobalEnv) after to extract the data
frames
}
\examples{
champsData <- read_rib("data-raw/champs.xlsx")
names(champsData)
champsData["KDA"]
}
